+++ docs/Changelog.md
@@ -12,6 +12,9 @@
 * R package
   * Support user-specified cross validation division for `abess` function by additional argument `foldid`. 
 
+* R package 
+  * Support generalized linear model when the link function is gamma distribution. It is help for the dataset with a positive valued and skewed response. 
+
 * Project development
   * Documentation
     * Add instruction for user-specified cross validation division.